third_party/WebKit/LayoutTests/media/video-enter-fullscreen-without-user-gesture.html

<html>
    <head>
        <title>Test webkitRequestFullscreen() without any user gesture.</title>
        <script src=media-file.js></script>
        <script src=video-test.js></script>
        <script>
            function canplaythrough()
            {
                consoleWrite("");
                consoleWrite("* No user gesture initiated");
                testExpected("video.paused", true);
                run("video.webkitRequestFullscreen()");
                waitForEvent('webkitfullscreenerror', endTest);
            }

            function start()
            {
                 findMediaElement();
                 waitForEvent('canplaythrough', canplaythrough);
                 video.src = findMediaFile("video", "content/test");
            }
        </script>
    </head>

    <body onload="start()">
    <p>Test webkitRequestFullscreen() without any user gesture.</p>
    <video controls></video>
    </body>
</html>
