third_party/WebKit/LayoutTests/media/video-enter-fullscreen-without-user-gesture-expected.txt

Test webkitRequestFullscreen() without any user gesture.

EVENT(canplaythrough)

* No user gesture initiated
EXPECTED (video.paused == 'true') OK
RUN(video.webkitRequestFullscreen())
EVENT(webkitfullscreenerror)
END OF TEST
